+++ front/src/pages/SignUp/index.tsx
@@ -1,46 +1,14 @@
-import React, { useState, useCallback } from 'react';
-import { Redirect } from 'react-router-dom';
+import React from 'react';
 import styled from 'styled-components';
 import logo from '../../images/logo.png';
-import AgeSelector from './AgeSelector';
-import GenderSelector from './GenderSelector';
-import SignupButton from './SingupButton';
+import UserInfoForm from '../../components/UserInfoForm';
 
-interface Props {
-  location: Location;
-}
-
-function SignUp({ location }: Props): JSX.Element {
-  const [nickname, setNickname] = useState('');
-  const [gender, setGender] = useState(0);
-  const [age, setAge] = useState(0);
-  const [authenticated, setAuthenticated] = useState(false);
-
-  const nicknameOnchange = useCallback((event: React.ChangeEvent<HTMLInputElement>) => {
-    const {
-      target: { value },
-    } = event;
-    setNickname(value);
-  }, []);
-
-  return authenticated ? (
-    <Redirect to="/main" />
-  ) : (
+function SignUp(): JSX.Element {
+  return (
     <Container>
       <img src={logo} alt="logo" width="220px" height="220px" />
       <Title>Welcome to world cup</Title>
-      <InputContainer>
-        <NameInput type="text" placeholder="닉네임" onChange={nicknameOnchange} />
-        <GenderSelector gender={gender} setGender={setGender} />
-        <AgeSelector age={age} setAge={setAge} />
-        <SignupButton
-          location={location}
-          nickname={nickname}
-          gender={gender}
-          age={age}
-          setAuthenticated={setAuthenticated}
-        />
-      </InputContainer>
+      <UserInfoForm />
     </Container>
   );
 }
@@ -61,24 +29,6 @@ const Title = styled.div`
   ${({ theme }) => theme.fontStyle.h1Bold}
   position : relative;
   top: -60px;
-`;
-
-const InputContainer = styled.form`
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  width: 389px;
-  height: 360px;
-`;
-
-const NameInput = styled.input`
-  ${({ theme }) => theme.fontStyle.h3}
-  background-color: ${({ theme }) => theme.color.primary};
-  padding-left: 32px;
-  width: 100%;
-  height: 61px;
-  border: 0;
-  border-radius: 10px;
 `;
 
 export default SignUp;
